requirements.txt
fastapi[standard]
pymilvus 
python-dotenv
sentence-transformers
pymilvus[model]
prometheus_client